+++ R/teams.R
@@ -78,14 +78,10 @@ setMethod("members", "CrunchTeam", function (x) {
 
 #' @rdname delete
 #' @export
-setMethod("delete", "CrunchTeam", function (x, confirm=requireConsent(), ...) {
-    if (!missing(confirm)) {
-        warning("The 'confirm' argument is deprecated. See ?with_consent.",
-            call.=FALSE)
-    }
+setMethod("delete", "CrunchTeam", function (x, ...) {
     prompt <- paste0("Really delete team ", dQuote(name(x)), "? ",
         "This cannot be undone.")
-    if (confirm && !askForPermission(prompt)) {
+    if (!askForPermission(prompt)) {
         halt("Must confirm deleting team")
     }
     u <- self(x)